+++ tasks/karma.js
@@ -35,12 +35,7 @@ module.exports = {
 
     // Continuous Integration mode
     // if true, Karma captures browsers, runs the tests and exits
-    singleRun: true,
-
-    // HTML templates mock
-    ngHtml2JsPreprocessor: {
-      moduleName: 'templates'
-    }
+    singleRun: true
 
   },